+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:example/example_regular.dart';
+
 import 'package:flutter/material.dart';
 
 void main() {