+++ lib/data/renderer_properties.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:light_html_editor/data/text_constants.dart';
 
+///
+/// collection of style properties for the RichtextRenderer
+///
 class RendererDecoration {
   final BoxBorder? border;
   final String? label;
@@ -9,6 +12,11 @@ class RendererDecoration {
   final double defaultFontSize;
   final EdgeInsets padding;
   final double? maxHeight;
+  final Color? linkColor;
+  final bool? linkUnderline;
+
+  /// displayed at the end of a shortened message due to length limits
+  final String overflowIndicator;
 
   const RendererDecoration({
     this.border,
@@ -18,5 +26,8 @@ class RendererDecoration {
     this.defaultFontSize = TextConstants.defaultFontSize,
     this.padding = const EdgeInsets.all(8),
     this.maxHeight,
+    this.linkColor,
+    this.linkUnderline,
+    this.overflowIndicator = "...",
   });
 }